backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-cloud-bigquery
google-auth
google-auth-httplib2
google-api-python-client
